Drop duplicate compose aws/credentials.js
Compose's lib/compose/utils/aws/credentials.js shared 122 lines of helpers
byte-for-byte with lib/aws/credentials.js. The only divergence was an
exported getCredentialProvider({ profile, stage }) — a stripped-down sibling
of the parent's getAwsSdkV3CredentialsProvider that omits the cliProfile and
providerProfile branches (compose has no Serverless provider object). Move
that 14-line function into lib/aws/credentials.js next to its sibling,
repoint compose's get-credential-provider.js, port the 6 stage-resolution
priority tests into the parent's test file, and delete the duplicate.
