+++ lib/aws/credentials.js
@@ -167,10 +167,28 @@ function getAwsSdkV3CredentialsProvider({ provider, profile } = {}) {
   return fromImplicitDefaultProfileWithFallback();
 }
 
+function getCredentialProvider({ profile, stage } = {}) {
+  const stageUpper = stage ? stage.toUpperCase() : null;
+
+  if (profile) return fromProfile(profile);
+  if (stageUpper && process.env[`AWS_${stageUpper}_PROFILE`]) {
+    return fromProfile(process.env[`AWS_${stageUpper}_PROFILE`]);
+  }
+  if (stageUpper && hasEnvironmentCredentials(`AWS_${stageUpper}`)) {
+    return fromPrefixedEnv(`AWS_${stageUpper}`);
+  }
+  if (process.env.AWS_PROFILE) return fromProfile(process.env.AWS_PROFILE);
+  if (hasEnvironmentCredentials('AWS')) return fromPrefixedEnv('AWS');
+  if (process.env.AWS_DEFAULT_PROFILE) return fromProfile(process.env.AWS_DEFAULT_PROFILE);
+
+  return fromImplicitDefaultProfileWithFallback();
+}
+
 module.exports = {
   doesImplicitDefaultProfileExist,
   fromPrefixedEnv,
   getAwsSdkV3CredentialsProviderCacheKey,
   getAwsSdkV3CredentialsProvider,
+  getCredentialProvider,
   hasEnvironmentCredentials,
 };
