+++ lib/compose/utils/aws/get-credential-provider.js
@@ -1,5 +1,5 @@
 'use strict';
 
-const { getCredentialProvider } = require('./credentials');
+const { getCredentialProvider } = require('../../../aws/credentials');
 
 module.exports = ({ profile, stage } = {}) => getCredentialProvider({ profile, stage });
